chore: Update dependencies and ESLint configuration
This commit updates the dependencies in the `package.json` file, adding `@react-router/dev` and `vite-tsconfig-paths` packages. It also modifies the ESLint configuration in the `.eslintrc.cjs` file, commenting out the `react-refresh/only-export-components` rule.

The changes in the `tsconfig.json` file include adding additional file patterns to the `include` section, allowing TypeScript to compile files in the `.server` and `.client` directories.

In the `vite.config.ts` file, the commit adds two new plugins: `cloudflareDevProxyVitePlugin` and `tsconfigPaths`. These plugins enhance the Vite configuration for better development experience and path resolution.

This commit is a chore because it includes updates and modifications that do not directly affect the functionality of the application.